Move PortAllocator into miles.utils.workers
