+++ tests/fast/ray/rollout/real_ray/test_server_cell.py
@@ -8,8 +8,8 @@ from tests.fast.ray.rollout.conftest import make_args
 from tests.fast.ray.rollout.real_ray.conftest import build_cells, start_cells
 
 from miles.backends.sglang_utils.sglang_engine import build_server_url
-from miles.ray.rollout.addr_allocator import PortAllocator
 from miles.ray.rollout.rollout_server import RolloutServer
+from miles.utils.workers.addr_allocator import PortAllocator
 
 
 def _all_actor_handles(cells) -> list:
